dragonpilot for 0.8.9-1

+++ launch_chffrplus.sh
@@ -9,6 +9,21 @@ source "$BASEDIR/launch_env.sh"
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 
 function two_init {
+  python /data/openpilot/scripts/installers/font_installer.py
+  python /data/openpilot/scripts/installers/sshkey_installer.py
+  mount -o remount,rw /system
+  if [ ! -f /ONEPLUS ] && ! $(grep -q "letv" /proc/cmdline); then
+    sed -i -e 's#/dev/input/event1#/dev/input/event2#g' ~/.bash_profile
+    touch /ONEPLUS
+  else
+    if [ ! -f /LEECO ]; then
+      touch /LEECO
+    fi
+  fi
+  mount -o remount,r /system
+
+  /data/data/com.termux/files/usr/bin/python /data/openpilot/scripts/installers/font_installer.py
+  /data/data/com.termux/files/usr/bin/python /data/openpilot/scripts/installers/sshkey_installer.py
 
   # set IO scheduler
   setprop sys.io.scheduler noop
@@ -66,6 +81,9 @@ function two_init {
 
   # USB traffic needs realtime handling on cpu 3
   [ -d "/proc/irq/733" ] && echo 3 > /proc/irq/733/smp_affinity_list
+  if [ -f /ONEPLUS ]; then
+    [ -d "/proc/irq/736" ] && echo 3 > /proc/irq/736/smp_affinity_list # USB for OP3T
+  fi
 
   # GPU and camera get cpu 2
   CAM_IRQS="177 178 179 180 181 182 183 184 185 186 192"
@@ -89,19 +107,33 @@ function two_init {
   wpa_cli IFNAME=wlan0 SCAN
 
   # Check for NEOS update
-  if [ $(< /VERSION) != "$REQUIRED_NEOS_VERSION" ]; then
-    if [ -f "$DIR/scripts/continue.sh" ]; then
-      cp "$DIR/scripts/continue.sh" "/data/data/com.termux/files/continue.sh"
-    fi
+  if [ -f /LEECO ]; then
+    if [ $(< /VERSION) != "$REQUIRED_NEOS_VERSION" ]; then
+      if [ -f "$DIR/scripts/continue.sh" ]; then
+        cp "$DIR/scripts/continue.sh" "/data/data/com.termux/files/continue.sh"
+      fi
 
-    if [ ! -f "$BASEDIR/prebuilt" ]; then
-      # Clean old build products, but preserve the scons cache
-      cd $DIR
-      git clean -xdf
-      git submodule foreach --recursive git clean -xdf
-    fi
+      if [ ! -f "$BASEDIR/prebuilt" ]; then
+        # Clean old build products, but preserve the scons cache
+        cd $DIR
+        git clean -xdf
+        git submodule foreach --recursive git clean -xdf
+      fi
 
-    "$DIR/installer/updater/updater" "file://$DIR/installer/updater/update.json"
+      "$DIR/installer/updater/updater" "file://$DIR/installer/updater/update.json"
+    fi
+  fi
+
+  # One-time fix for a subset of OP3T with gyro orientation offsets.
+  # Remove and regenerate qcom sensor registry. Only done on OP3T mainboards.
+  # Performed exactly once. The old registry is preserved just-in-case, and
+  # doubles as a flag denoting we've already done the reset.
+  if [ -f /ONEPLUS ] && [ ! -f "/persist/comma/op3t-sns-reg-backup" ]; then
+    echo "Performing OP3T sensor registry reset"
+    mv /persist/sensors/sns.reg /persist/comma/op3t-sns-reg-backup &&
+      rm -f /persist/sensors/sensors_settings /persist/sensors/error_log /persist/sensors/gyro_sensitity_cal &&
+      echo "restart" > /sys/kernel/debug/msm_subsys/slpi &&
+      sleep 5  # Give Android sensor subsystem a moment to recover
   fi
 }
 
@@ -152,10 +184,10 @@ function launch {
   #    that completed successfully and synced to disk.
 
   if [ -f "${BASEDIR}/.overlay_init" ]; then
-    find ${BASEDIR}/.git -newer ${BASEDIR}/.overlay_init | grep -q '.' 2> /dev/null
-    if [ $? -eq 0 ]; then
-      echo "${BASEDIR} has been modified, skipping overlay update installation"
-    else
+#    find ${BASEDIR}/.git -newer ${BASEDIR}/.overlay_init | grep -q '.' 2> /dev/null
+#    if [ $? -eq 0 ]; then
+#      echo "${BASEDIR} has been modified, skipping overlay update installation"
+#    else
       if [ -f "${STAGING_ROOT}/finalized/.overlay_consistent" ]; then
         if [ ! -d /data/safe_staging/old_openpilot ]; then
           echo "Valid overlay update found, installing"
@@ -179,13 +211,20 @@ function launch {
           # TODO: restore backup? This means the updater didn't start after swapping
         fi
       fi
-    fi
+#    fi
   fi
 
   # handle pythonpath
   ln -sfn $(pwd) /data/pythonpath
   export PYTHONPATH="$PWD:$PWD/pyextra"
 
+  # dp - ignore chmod changes
+  git -C $DIR config core.fileMode false
+
+  # dp - apply custom patch
+  if [ -f "/data/media/0/dp_patcher.py" ]; then
+    python /data/media/0/dp_patcher.py
+  fi
   # hardware specific init
   if [ -f /EON ]; then
     two_init
@@ -198,7 +237,21 @@ function launch {
 
   # start manager
   cd selfdrive/manager
-  ./build.py && ./manager.py
+  if [ -f /EON ]; then
+    if [ ! -f "/system/comma/usr/lib/libgfortran.so.5.0.0" ]; then
+      mount -o remount,rw /system
+      tar -zxvf /data/openpilot/selfdrive/mapd/assets/libgfortran.tar.gz -C /system/comma/usr/lib/
+      mount -o remount,r /system
+    fi
+    if [ ! -d "/system/comma/usr/lib/python3.8/site-packages/opspline" ]; then
+      mount -o remount,rw /system
+      tar -zxvf /data/openpilot/selfdrive/mapd/assets/opspline.tar.gz -C /system/comma/usr/lib/python3.8/site-packages/
+      mount -o remount,r /system
+    fi
+    ./build.py && ./manager.py
+  else
+    ./custom_dep.py && ./build.py && ./manager.py
+  fi
 
   # if broken, keep on screen error
   while true; do sleep 1; done
